requirements.txt
Django
Pillow
djangorestframework
django-filter
